på wcf siden tager den fra databsen istedet for to bestemte tekst strenge
