+++ Projekt Mappe/DrinkzyWCF/WCF/AuthService.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using BusinessLayer;
+using ModelLayer;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -8,9 +10,17 @@ namespace WCF
 {
     public class AuthService : IAuthService
     {
+        UserController uCtr = new UserController();
         public bool Login(string username, string password)
         {
-            return "SuperStudent" == username && "1234" == password;
+            User user = uCtr.GetUserByUserName(username, password);
+            Boolean auth;
+            if (user != null)
+            {
+                auth = true;
+            }
+            else { auth = false; }
+            return auth;
         }
     }
 }
