+++ Projekt Mappe/Website/Gui/Controllers/HomeController.cs
@@ -18,16 +18,6 @@ namespace Gui.Controllers
        
         public ActionResult Index()
         {
-            ServicePointManager.ServerCertificateValidationCallback = (obj, certificate, chain, errors) => true;
-            AuthServiceClient authClient = new AuthServiceClient();
-            var isLoggedIn = authClient.Login("SuperStudent", "1234");
-            if (isLoggedIn)
-            {
-                SecurityServiceClient client = new SecurityServiceClient("WSHttpBinding_ISecurityService");
-                client.ClientCredentials.UserName.UserName = "SuperStudent";
-                client.ClientCredentials.UserName.Password = "1234";
-                var data = client.GetData(1337);
-            }
             return View();
         }
