+++ Projekt Mappe/Website/Gui/Views/Shared/_Layout.cshtml
@@ -43,7 +43,7 @@
                                     <div class="col-md-12">
                                         <form class="form" role="form" method="post" action="login" accept-charset="UTF-8" id="login-nav">
                                             <div class="form-group">
-                                                <label class="sr-only" for="exampleInputEmail2">Email address</label>
+                                                <label class="sr-only" for="exampleInputEmail2">User Name</label>
                                                 <input type="email" class="form-control" id="exampleInputEmail2" placeholder="Email address" required>
                                             </div>
                                             <div class="form-group">
